requirements.txt
discord.py     >= 1.7.3
git+https://github.com/nineclicks/trivia_core.git